create utils file, move run_subprocess into utils
